requirements.txt
flask==3.0.3
flask-cors==4.0.1
langchain==0.2.16
langchain-groq==0.1.23
langgraph==0.2.28
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
